Move MHA with PE tests to a new test file
